--- src/main/java/codesquad/fineants/domain/kis/properties/KisCodeValue.java
@@ -1,5 +1,0 @@
-package codesquad.fineants.domain.kis.properties;
-
-public interface KisCodeValue {
-	String getCode();
-}
