+++ src/main/java/codesquad/fineants/domain/kis/properties/kiscodevalue/KisCodeValue.java
@@ -1,0 +1,5 @@
+package codesquad.fineants.domain.kis.properties.kiscodevalue;
+
+public interface KisCodeValue {
+	String getCode();
+}
